Move start button into separate module
